master：以图搜款功能完善；

+++ xkt/src/main/java/com/ruoyi/xkt/dto/picture/ProductImgSearchCountDTO.java
@@ -17,5 +17,5 @@ public class ProductImgSearchCountDTO implements Serializable {
     /**
      * 图搜次数
      */
-    private Long imgSearchCount;
+    private Integer imgSearchCount;
 }
